# docs/architecture/openapi-baseline.json
{
 "components": {
  "schemas": {
...
   "ContactMethod": {
    "description": "How the customer wants sales to follow up.",
    "enum": [
     "call",
     "agent",
     "whatsapp"
    ],
    "title": "ContactMethod",
    "type": "string"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "LoginIn": {
    "properties": {
     "password": {
      "title": "Password",
      "type": "string"
     },
     "username": {
      "title": "Username",
      "type": "string"
     }
    },
    "required": [
     "username",
     "password"
    ],
    "title": "LoginIn",
    "type": "object"
   },
   "MeOut": {
    "properties": {
     "role": {
      "default": "superadmin",
      "title": "Role",
      "type": "string"
     },
     "username": {
      "title": "Username",
      "type": "string"
     }
    },
    "required": [
     "username"
    ],
    "title": "MeOut",
    "type": "object"
   },
   "OrderCreate": {
    "properties": {
     "city": {
      "anyOf": [
       {
        "maxLength": 60,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "City"
     },
     "contact_method": {
      "$ref": "#/components/schemas/ContactMethod",
      "default": "call"
     },
     "full_name": {
      "maxLength": 60,
      "minLength": 3,
      "title": "Full Name",
      "type": "string"
     },
     "items": {
      "items": {
       "$ref": "#/components/schemas/OrderItemIn"
      },
      "minItems": 1,
      "title": "Items",
      "type": "array"
     },
     "note": {
      "anyOf": [
       {
        "maxLength": 300,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     },
     "phone": {
      "pattern": "^09\\d{9}$",
      "title": "Phone",
      "type": "string"
     },
     "province": {
      "title": "Province",
      "type": "string"
     },
     "referrer": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Referrer"
     },
     "store_name": {
      "maxLength": 80,
      "minLength": 2,
      "title": "Store Name",
      "type": "string"
     },
     "utm_campaign": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Utm Campaign"
     },
     "utm_medium": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Utm Medium"
     },
     "utm_source": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Utm Source"
     },
     "website": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Website"
     }
    },
    "required": [
     "full_name",
     "phone",
     "store_name",
     "province",
     "items"
    ],
    "title": "OrderCreate",
    "type": "object"
   },
   "OrderCreatedOut": {
    "properties": {
     "reference": {
      "title": "Reference",
      "type": "string"
     },
     "total": {
      "title": "Total",
      "type": "string"
     }
    },
    "required": [
     "reference",
     "total"
    ],
    "title": "OrderCreatedOut",
    "type": "object"
   },
...
   "OrderItemIn": {
    "properties": {
     "product_id": {
      "anyOf": [
       {
        "format": "uuid",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Product Id"
     },
     "quantity": {
      "maximum": 999.0,
      "minimum": 1.0,
      "title": "Quantity",
      "type": "integer"
     }
    },
    "required": [
     "quantity"
    ],
    "title": "OrderItemIn",
    "type": "object"
   },
   "OrderItemOut": {
    "properties": {
     "product_id": {
      "anyOf": [
       {
        "format": "uuid",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Product Id"
     },
     "product_name": {
      "title": "Product Name",
      "type": "string"
     },
     "quantity": {
      "title": "Quantity",
      "type": "integer"
     },
     "unit_price": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Unit Price"
     },
     "unit_weight_grams": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Unit Weight Grams"
     }
    },
    "required": [
     "product_id",
     "product_name",
     "unit_weight_grams",
     "unit_price",
     "quantity"
    ],
    "title": "OrderItemOut",
    "type": "object"
   },
...
   "OrderStatus": {
    "description": "WO 7.4 mapping: new≈Submitted, confirmed≈Approved, shipped≈Dispatched.",
    "enum": [
     "new",
     "contacted",
     "confirmed",
     "preparing",
     "shipped",
     "delivered",
     "cancelled"
    ],
    "title": "OrderStatus",
    "type": "string"
   },
   "OrderStatusLogOut": {
    "properties": {
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "from_status": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/OrderStatus"
       },
       {
        "type": "null"
       }
      ]
     },
     "to_status": {
      "$ref": "#/components/schemas/OrderStatus"
     }
    },
    "required": [
     "from_status",
     "to_status",
     "created_at"
    ],
    "title": "OrderStatusLogOut",
    "type": "object"
   },
   "OrderTrackOut": {
    "description": "Public order-status view (reference + phone lookup). Deliberately omits\nstored PII (name/address) — the customer already has it; the response only\nconfirms status + line items + the status timeline.",
    "properties": {
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "items": {
      "items": {
       "$ref": "#/components/schemas/OrderItemOut"
      },
      "title": "Items",
      "type": "array"
     },
     "reference": {
      "title": "Reference",
      "type": "string"
     },
     "serial_codes": {
      "default": [],
      "items": {
       "type": "string"
      },
      "title": "Serial Codes",
      "type": "array"
     },
     "status": {
      "$ref": "#/components/schemas/OrderStatus"
     },
     "status_log": {
      "items": {
       "$ref": "#/components/schemas/OrderStatusLogOut"
      },
      "title": "Status Log",
      "type": "array"
     },
     "total": {
      "title": "Total",
      "type": "string"
     }
    },
    "required": [
     "reference",
     "status",
     "total",
     "created_at",
     "items",
     "status_log"
    ],
    "title": "OrderTrackOut",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "Didar Gold API",
  "version": "1.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/admin/login": {
   "post": {
    "description": "Named users first; the env-var admin stays as a zero-config bootstrap\nsuperadmin. Both outcomes are audited (login is the one mutation the audit\nmiddleware can't attribute — there's no cookie yet).",
    "operationId": "login_api_v1_admin_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LoginIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MeOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Login",
    "tags": [
     "auth"
    ]
   }
  },
...
  "/api/v1/orders": {
   "post": {
    "operationId": "create_order_api_v1_orders_post",
    "parameters": [
     {
      "in": "header",
      "name": "Idempotency-Key",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OrderCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OrderCreatedOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Order",
    "tags": [
     "public"
    ]
   }
  },
  "/api/v1/orders/track": {
   "get": {
    "description": "Account-less order tracking. Both reference AND the ordering phone must\nmatch — an unknown reference and a wrong phone return the SAME 404 so a\nreference can't be enumerated without the phone. Never cached (status changes).",
    "operationId": "track_order_api_v1_orders_track_get",
    "parameters": [
     {
      "in": "query",
      "name": "reference",
      "required": true,
      "schema": {
       "title": "Reference",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "phone",
      "required": true,
      "schema": {
       "title": "Phone",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OrderTrackOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Track Order",
    "tags": [
     "public"
    ]
   }
  },
...
 }
}